pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nba-props-ml"
version = "1.0.0"
description = "NBA player props ML prediction system with multi-book line shopping"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "untitled114", email = "untitled114@users.noreply.github.com"}
]
keywords = [
    "machine-learning",
    "sports-betting",
    "nba",
    "lightgbm",
    "predictions",
    "data-engineering"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "lightgbm>=4.0.0,<5.0.0",
    "scikit-learn>=1.3.0,<2.0.0",
    "numpy>=1.24.0,<3.0.0",
    "pandas>=2.0.0,<3.0.0",
    "scipy>=1.10.0,<2.0.0",
    "joblib>=1.3.0,<2.0.0",
    "psycopg2-binary>=2.9.0,<3.0.0",
    "pymongo>=4.0.0,<5.0.0",
    "nba_api>=1.4.0,<2.0.0",
    "beautifulsoup4>=4.12.0,<5.0.0",
    "playwright>=1.40.0,<2.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "pytz>=2024.1",
    "requests>=2.28.0,<3.0.0",
    "tqdm>=4.65.0,<5.0.0",
    "pydantic>=2.5.0,<3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "isort>=5.13.0",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]
viz = [
    "matplotlib>=3.7.0,<4.0.0",
    "seaborn>=0.12.0,<1.0.0",
]
mlops = [
    "mlflow>=2.9.0,<3.0.0",
]

[project.urls]
Repository = "https://github.com/untitled114/nba-props-betting"

[tool.setuptools.packages.find]
where = ["."]
include = ["nba*", "core*"]

# Black configuration
[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
    | \.venv
    | venv
    | __pycache__
    | \.eggs
    | build
    | dist
    | docker
    | \.mypy_cache
)/
'''

# isort configuration
[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true
skip = [".git", ".venv", "venv", "__pycache__", "docker"]
known_first_party = ["nba", "core"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

# pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
pythonpath = [".", "nba"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--cov=nba",
    "--cov-report=term-missing",
    "--cov-report=html:coverage_html",
    "--cov-report=xml:coverage.xml",
    # Coverage threshold disabled - external integration code excluded
    # Core library coverage tracked separately (target: 80%)
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
    "ignore::PendingDeprecationWarning",
]
markers = [
    "unit: Unit tests (fast, no external dependencies)",
    "integration: Integration tests (may need mocked services)",
    "slow: Slow tests (database queries, API calls)",
]

# mypy configuration
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = [
    "venv",
    ".venv",
    "docker",
    "build",
    "dist",
]

# Coverage configuration
[tool.coverage.run]
source = ["nba"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
    "*/migrations/*",
    # External integration code (requires DB/HTTP mocking for full coverage)
    "nba/scripts/*",                   # CLI scripts that call main()
    "nba/betting_xl/loaders/*",        # DB loaders
    "nba/betting_xl/fetchers/fetch_*", # HTTP fetchers
    "nba/betting_xl/validate_*",       # Validation scripts
    "nba/betting_xl/generate_*",       # Generation scripts
    "nba/betting_xl/run_*",            # Runner scripts
    "nba/features/build_*",            # Dataset builders
    "nba/models/train_*.py",           # Training scripts
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "except ImportError",
    "except Exception as e:",
    "logger\\.(debug|info|warning|error)",
    "def main\\(",
]
# Core library code target: 80%
# Overall (with omissions): ~40%
fail_under = 0  # Disabled - track manually until integration tests added
show_missing = true
skip_covered = false

[tool.coverage.html]
directory = "coverage_html"

[tool.coverage.xml]
output = "coverage.xml"

# Bandit security linter configuration
[tool.bandit]
exclude_dirs = ["tests", "venv", ".venv", "docker"]
skips = [
    "B101",  # assert (used in tests)
    "B110",  # try_except_pass (intentional in data parsing)
    "B112",  # try_except_continue (intentional in data parsing)
    "B311",  # random (not used for crypto, just user agents/backoff)
    "B301",  # pickle (required for ML model serialization)
    "B403",  # pickle import (required for ML)
    "B608",  # SQL injection (false positives - uses parameterized queries)
]
